+++ example/lib/main.dart
@@ -24,51 +24,6 @@ class MainApp extends StatelessWidget {
                   width: 100,
                   color: 'bada55'.toColor,
                 ),
-                const Gap(16),
-                FlexRow(
-                  gap: 16,
-                  mainAxisSize: MainAxisSize.min,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Container(
-                      height: 100,
-                      width: 100,
-                      color: 'bada55'.toColor,
-                    ),
-                    Container(
-                      height: 100,
-                      width: 100,
-                      color: '5eabed'.toColor,
-                    ),
-                    Container(
-                      height: 100,
-                      width: 100,
-                      color: 'facade'.toColor,
-                    ),
-                  ],
-                ),
-                const Gap(16),
-                FlexColumn(
-                  gap: 16,
-                  // reversed: true,
-                  children: [
-                    Container(
-                      height: 100,
-                      width: 100,
-                      color: 'bada55'.toColor,
-                    ),
-                    Container(
-                      height: 100,
-                      width: 100,
-                      color: '5eabed'.toColor,
-                    ),
-                    Container(
-                      height: 100,
-                      width: 100,
-                      color: 'facade'.toColor,
-                    ),
-                  ],
-                ),
               ],
             ),
           ),