+++ lib/src/extensions/widget.dart
@@ -14,16 +14,15 @@ extension WidgetExtension on Widget {
   ///
   /// ```dart
   /// Text('Hello, World!').toExpanded()
-  /// Text('Hello, World!').toExpanded(1)
+  /// Text('Hello, World!').toExpanded(flex: 1)
   /// ```
-  Widget toExpanded(int? n) => Expanded(flex: n ?? 1, child: this);
+  Widget toExpanded({int? flex}) => Expanded(flex: flex ?? 1, child: this);
 
   /// Converts a child of a [Row], [Column], or [Flex] widget to a [Flexible].
   ///
   /// ```dart
   /// Text('Hello, World!').toFlexible()
-  /// Text('Hello, World!').toFlexible(1, fit: FlexFit.loose)
+  /// Text('Hello, World!').toFlexible(flex: 1, fit: FlexFit.loose)
   /// ```
-  Widget toFlexible(int? n, {FlexFit? fit}) =>
-      Flexible(flex: n ?? 1, fit: fit ?? FlexFit.loose, child: this);
+  Widget toFlexible({int? flex, FlexFit? fit}) => Flexible(flex: flex ?? 1, fit: fit ?? FlexFit.loose, child: this);
 }