+++ lib/src/widgets/widgets.dart
@@ -1,2 +1,1 @@
 export 'gap.dart';
-export 'flexbox.dart';